+++ Tests/Integration/MSFT_PowerShellExecutionPolicy.config.ps1
@@ -1,0 +1,14 @@
+# Integration Test Config Template Version: 1.0.0
+configuration MSFT_PowerShellExecutionPolicy_config
+ {
+    Import-DscResource -ModuleName ComputerManagementDsc
+
+    node 'localhost'
+    {
+        PowerShellExecutionPolicy Integration_Test
+        {
+            ExecutionPolicy      = 'RemoteSigned'
+            ExecutionPolicyScope = 'LocalMachine'
+        }
+    }
+}